feat: extract reusable capture crate
