+++ agentsight-capture/src/analyzers/auth_header_remover.rs
