+++ agentsight-capture/src/analyzers/filter_base.rs
