+++ agentsight-capture/src/analyzers/filter_metrics.rs
