+++ agentsight-capture/src/analyzers/materializing.rs
