+++ agentsight-capture/src/analyzers/mod.rs
