+++ agentsight-capture/src/analyzers/mod_tests.rs
