+++ agentsight-capture/src/analyzers/sse_processor_tests.rs
