+++ agentsight-capture/src/binary_extractor.rs
