+++ agentsight-capture/src/json.rs
