+++ agentsight-capture/src/runners/common.rs
