+++ agentsight-capture/src/sinks/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 eunomia-bpf org.
+
+pub mod otel;
+pub mod sqlite;
+
+pub use otel::OtelExporter;
