+++ agentsight-capture/src/sources/mod.rs
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2026 eunomia-bpf org.
 
-pub mod otel;
-pub(crate) mod sqlite;
-
-pub use otel::OtelExporter;
+pub mod agent_native;
+pub mod proc;
+pub mod sqlite;
