+++ agentsight-capture/src/view/canonical.rs
