+++ collector/build.rs
@@ -8,7 +8,6 @@ fn main() {
     let repo_root = manifest_dir.parent().unwrap_or(&manifest_dir);
     let sync_vendor = env_flag("AGENTSIGHT_SYNC_VENDOR");
 
-    sync_bpf(&manifest_dir, repo_root, sync_vendor);
     sync_frontend(&manifest_dir, repo_root, sync_vendor);
 
     println!("cargo:rerun-if-changed=build.rs");
@@ -19,38 +18,6 @@ fn env_flag(name: &str) -> bool {
     env::var(name)
         .map(|value| matches!(value.as_str(), "1" | "true" | "yes" | "on"))
         .unwrap_or(false)
-}
-
-fn sync_bpf(manifest_dir: &Path, repo_root: &Path, sync_vendor: bool) {
-    let source_dir = repo_root.join("bpf");
-    let vendor_dir = manifest_dir.join("vendor/bpf");
-    for name in ["process", "sslsniff", "stdiocap"] {
-        let source = source_dir.join(name);
-        let vendor = vendor_dir.join(name);
-        if sync_vendor {
-            if !source.exists() {
-                panic!(
-                    "missing BPF loader {}. Run `make -C ../bpf` before packaging.",
-                    source.display()
-                );
-            }
-            println!("cargo:rerun-if-changed={}", source.display());
-            copy_file(&source, &vendor).unwrap_or_else(|err| {
-                panic!(
-                    "failed to vendor {} into {}: {err}",
-                    source.display(),
-                    vendor.display()
-                )
-            });
-        }
-        if !vendor.exists() {
-            panic!(
-                "missing bundled BPF loader {}. Run `make build` before packaging.",
-                vendor.display()
-            );
-        }
-        println!("cargo:rerun-if-changed={}", vendor.display());
-    }
 }
 
 fn sync_frontend(manifest_dir: &Path, repo_root: &Path, sync_vendor: bool) {
