+++ collector/src/sources/mod.rs
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2026 eunomia-bpf org.
 
-pub mod agent_native;
-pub mod proc;
+pub(crate) use agentsight_capture::sources::{agent_native, proc, sqlite};
+
 pub(crate) mod session_db;
-pub mod sqlite;
